requirements.txt
requests==2.31.0
openpyxl==3.1.2
python-telegram-bot==13.15
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
pandas==2.0.3
pytz==2023.3
schedule==1.2.0